% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allison_plot.R
\name{allison_plot}
\alias{allison_plot}
\title{Plot sonar + hamachan estimates for 2020 analysis}
\usage{
allison_plot(df, y_up, main)
}
\arguments{
\item{df}{data.frame with columns named est, sd, method}

\item{y_up}{(numeric) upper bound of y-axis}

\item{main}{(character) plot title}
}
\description{
Plot sonar + hamachan estimates for 2020 analysis
}
\author{
Matt Tyres and Jordy Bernard.
}
