% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_totals.R
\name{calculate_totals}
\alias{calculate_totals}
\title{Calculate totals, according to a user-input date range.}
\usage{
calculate_totals(final_results, from, to)
}
\arguments{
\item{final_results}{The output of extract_final_results.}

\item{from}{The start date (e.g. "2019-06-23").}

\item{to}{The end date (e.g. "2019-08-10").}
}
\description{
Calculate totals, according to a user-input date range.
}
\author{
Matt Tyres and Jordy Bernard
}
