% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_estimates.R
\name{compile_estimates}
\alias{compile_estimates}
\title{Compile visual and sonar estimates into a single data.frame.}
\usage{
compile_estimates(vis_expanded, son_expanded)
}
\arguments{
\item{vis_expanded}{Expanded visual counts output from \code{\link{expand_counts}}.}

\item{son_expanded}{Expanded sonar counts output from \code{\link{expand_counts}}.}
}
\description{
Compile visual and sonar estimates into a single data.frame.
}
\author{
Matt Tyres and Jordy Bernard.
}
