% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_final_results.R
\name{extract_final_results}
\alias{extract_final_results}
\title{Extract the final results from the model.}
\usage{
extract_final_results(hamachan_samples, historical_counts, ra_ests)
}
\arguments{
\item{hamachan_samples}{The output of \code{\link{run_hamachan}}.}

\item{historical_counts}{The output of \code{\link{prepare_historical_counts}}.}

\item{ra_ests}{The output of \code{\link{runningavg_interp}}.}
}
\description{
Extract the final results from the model.
}
\author{
Matt Tyres and Jordy Bernard.
}
