% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_convergence_issues.R
\name{find_convergence_issues}
\alias{find_convergence_issues}
\title{Identify parameters with Rhat values > 1.1}
\usage{
find_convergence_issues(mixture_samples)
}
\arguments{
\item{mixture_samples}{Output of \code{\link{run_mixture_model}}.}
}
\description{
Identify parameters with Rhat values > 1.1
}
\author{
Jordy Bernard.
}
