% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_length_data.R
\name{prepare_length_data}
\alias{prepare_length_data}
\title{Reads in carcass length data.}
\usage{
prepare_length_data(dir)
}
\arguments{
\item{dir}{The address of the directory where the carcass length data is stored.}
}
\description{
Reads in and formats carcass length data for use in \code{\link{get_length_samples}}.
}
\author{
Matt Tyres and Jordy Bernard.
}
