% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sonar_counts_for_expansion2.R
\name{prepare_sonar_counts_for_expansion_2}
\alias{prepare_sonar_counts_for_expansion_2}
\title{Formats mixture model output and visual count data for expansion.}
\usage{
prepare_sonar_counts_for_expansion_2(
  visual_counts,
  sonar_counts,
  mixture_samples,
  trunc = 450,
  lthresh = 650
)
}
\arguments{
\item{visual_counts}{Visual count data. See main_2019.R for more information.}

\item{sonar_counts}{Sonar count data. See main_2019.R for more information.}

\item{mixture_samples}{Output of \code{\link{run_mixture_model}}.}

\item{trunc}{Lower bound for post-truncation based on input length. In past years, 450 seemed to work best.}
}
\description{
The posterior totals are formatted as a set of arrays, in which each cell corresponds to a 20-min counting block.  Array dimensions correspond to date, 8-hr shift, hour period, and 20-min block. Arrays are generated for point estimates (median) and variance, for each species and each sonar station. This information will be taken as input in \code{\link{expand_counts}}.
}
\author{
Matt Tyres and Jordy Bernard.
}
