% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_visual_counts_for_expansion.R
\name{prepare_visual_counts_for_expansion}
\alias{prepare_visual_counts_for_expansion}
\title{Format the visual counts for expansion}
\usage{
prepare_visual_counts_for_expansion(visual_counts, sonar_expansion_data)
}
\arguments{
\item{visual_counts}{Visual count data. See examples.}

\item{sonar_expansion_data}{Output of prepare_sonar_counts_for_expansion.}
}
\description{
Function formats the visual counts so that they can be expanded using the function expand_counts. This function will return an array where each cell corresponds to a 20-min counting block.  Array dimensions correspond to date, 8-hr shift, hour period, and 20-min block.
}
