% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFile_ARIS.R
\name{ReadFile_ARIS}
\alias{ReadFile_ARIS}
\title{Reads ARIS files. Called internally within ProcessFiles.}
\usage{
ReadFile_ARIS(filename = "tempfile.txt")
}
\arguments{
\item{filename}{File to be processed}
}
\description{
Returns a data frame containing the rows from the file columns for date, start time, the end time, and duration.
}
\author{
Allison Matter and Carl Pfisterer
}
