% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hamachan.R
\name{run_hamachan}
\alias{run_hamachan}
\title{Run Hamachan's hierarchical run-timing model.}
\usage{
run_hamachan(historical_counts, file_dir, ncores = 3, niter = 2e+05)
}
\arguments{
\item{historical_counts}{Output of \code{\link{prepare_historical_counts}}.}

\item{file_dir}{The file directory to write the jags model. The directory needs to have a .jags extension.}

\item{ncores}{The number of cores for parallel chains.}

\item{niter}{The number of MCMC iterations.}
}
\description{
Run Hamachan's hierarchical run-timing model.
}
\author{
Jordy Bernard.
}
