% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_sonar_plot.R
\name{visual_sonar_plot}
\alias{visual_sonar_plot}
\title{Plot expanded visual and sonar counts.}
\usage{
visual_sonar_plot(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{compile_estimates}}.}
}
\description{
Plot expanded visual and sonar counts.
}
\author{
Matt Tyres and Jordy Bernard.
}
